requirements.txt
asgiref==3.8.1
betterpath==0.2.2
click==8.1.7
Django==5.0.6
django-environ==0.11.2
django-filter==24.2
djangorestframework==3.15.1
l==0.11.0
Markdown==3.6
psycopg2-binary==2.9.9
setuptools==70.0.0
sqlparse==0.5.0
vcversioner==2.16.0.0
zope.interface==6.4.post2
